Start profiling for performance bottlenecks
The most blatant performance issue was the use of std::vector in the
ConservationLaw::State struct. To a much smaller extent, Cow::Matrix
is also an issue. I got a 3x speedup by switching from std::vector to
statically allocated std::array. This means the code will be compiled
with support for some maximum number of conserved quantities, which
is totally reasonable. Next thing to work out in terms of performance
is eliminating redundant calculations in the fromPrimitive function.
Left and right eigenvector matrices should perhaps be computed in a
separate function rather than being part of the State struct.