Add BC methods to support CT, and compute div-B using them
I feel that to extend the flux arrays in the solver would break
encapsulation, by exposing to CT the number of guard zones utilized by
the solver, and exposing to the solver the number of guard zones
required by CT. Exposing the BC to both modules, is more natural, and
ensures that they both receive and export data in a manner that is
agnostic to requirements of other modules. The additional
responsibility falls the BC class, which now must be able to apply the
boundary condition to more general types of data: B's as well as P's,
and data stored on different parts of the mesh.

Right now, the CT module is specialized to 2D.