Think more about dependency injection
I'm starting to think that abstracting the supply of
dependencies is not necessarily a good idea. The point would
have been for derived component classes to indicate somehow that
they require a particular service, and for the initialization
stage to coordinate supplying those services automatically. But,
given the relatively small number of interdependencies, it seems
more reasonable for the initialization code to simply be aware of
which component classes require a given service, in general. Whether
the derived class actually uses that service (by overriding the set
corresponding method) does not need to be known by the initialization
code.