Write MeshOpeator::curl and test return size
The MeshOperator is more specifically a "finite-area / finite-volume
operations dispatcher". I'm not sure how to name this thing. Really
it evaluates Stoke's theorem on faces and volumes. So maybe it should
be called DifferentialForm. It also is responsible for generating
new differential forms and their integrals over finite patches of the
mesh.

It's also not clear how the implementation for logically cartesian
meshes should interface with MeshGeometry instances in the future,
since the "DifferentialForm" requires knowledge of the MeshGeometry
index space; i.e. it knows the meaning of i,j,k. If later on, i,j,k
indicate locations in a hex-mesh for instance, then the
HexMeshGeometry would need to pair one-to-one with HexDifferentialForm.
So, maybe both classs need to be combined together in a Mesh class.