Almost finish with refactoring FluxConservativeSystem
Before, this class was getting out of control. After looking
at its methods and refactoring, I found its responsibilities
belong in four separate classes:

- MeshData       -> nothing : store and retrieve primitive data
- MeshOperator   -> MeshGeometry : finite-volume ops, generate data
- FieldOperator  -> ConservationLaw : point-wise ops, diagnostics
- SolutionScheme -> All of the above : TVD, BC's, etc

MeshData is responsible only for storing and retrieving current
solution data. It does not depend on any physics modules. However
it does keep tabs on which sections of data are internal, and
which are guard zone regions. Queries can get Array references
to the physical data. SolutionScheme dictates the number of
guard zones it requires, which is used to allocate MeshData.
SolutionScheme operates on MeshData objects, and maintains
MeshOperator, FieldOperator, and BoundaryCondition as
dependencies.