Fix bug in RK sink locations... LIVE BINARY DISABLED
When implementing the live binary, the RK steps in SolutionScheme were
written to accommodate sink particle derivatives. The steps are correct
if the derivatives are being used (live binary case), but wrong if the
sinkParticleLocations are provided as an explicit function of time. The
source terms function was being called with the same locations for both
of the RK steps.